NOT COMPILABLE Created abstracs Querys
