+++ QTFK.DB/QTFK.DB.csproj
@@ -67,6 +67,10 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\DBIO\AbstractDBIO.cs" />
     <Compile Include="Services\DBIO\AbstractQueryFactory.cs" />
+    <Compile Include="Models\DBIO\AbstractDeleteQuery.cs" />
+    <Compile Include="Models\DBIO\AbstractInsertQuery.cs" />
+    <Compile Include="Models\DBIO\AbstractSelectQuery.cs" />
+    <Compile Include="Models\DBIO\AbstractUpdateQuery.cs" />
     <Compile Include="Services\DBIO\TransactSQLParameterBuilder.cs" />
     <Compile Include="Services\IDBIO.cs" />
     <Compile Include="Services\IParameterBuilder.cs" />
