+++ QTFK.Services.DBIO.SQLServer/Services/DBIO/SqlServerQueryFactory.cs
@@ -14,11 +14,11 @@ namespace QTFK.Services.DBIO
             return new SQLServerQueryFactory(new Type[]
             {
                 //typeof(SqlByParamEqualsFilter),
-            });
+            }, new TransactSQLParameterBuilder());
         }
 
-        public SQLServerQueryFactory(IEnumerable<Type> filterTypes) 
-            : base(filterTypes)
+        public SQLServerQueryFactory(IEnumerable<Type> filterTypes, IParameterBuilder parameterBuilder) 
+            : base(filterTypes, parameterBuilder)
         {
         }
 
@@ -29,7 +29,7 @@ namespace QTFK.Services.DBIO
 
         protected override IDBQueryInsert prv_newInsert()
         {
-            return new SqlInsertQuery();
+            return new SqlInsertQuery(this.parameterBuilder);
         }
 
         protected override IDBQuerySelect prv_newSelect()
@@ -39,7 +39,7 @@ namespace QTFK.Services.DBIO
 
         protected override IDBQueryUpdate prv_newUpdate()
         {
-            return new SqlUpdateQuery();
+            return new SqlUpdateQuery(this.parameterBuilder);
         }
     }
 }
