+++ Tests/QTFK.Services.DBIO.OleDB.Tests/OleDBTests.cs
@@ -9,7 +9,6 @@ using QTFK.Extensions.DBIO.OleDBIOExtensions;
 using QTFK.Extensions.Mapping.AutoMapping;
 using QTFK.Models;
 using QTFK.Models.DBIO;
-using QTFK.Models.DBIO.Filters;
 using QTFK.Services.DBIO.OleDB.Tests.Models;
 using System;
 using System.Collections.Generic;
